# hedge-swagger-ui/res/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "hEdge APIs",
        "contact": {},
        "version": "v3"
    },
    "host": "localhost:48100",
    "basePath": "/hedge/",
    "paths": {
        "/api/v3/dataGenerator/commands": {
            "put": {
                "description": "You might change the average, or min, max etc depending on what you want.",
                "tags": [
                    "hedge-device-virtual - simulated data generator"
                ],
                "summary": "updates the seed data generation parameters so subsequent data values are based on new seed",
                "parameters": [
                    {
                        "description": "array of commands.",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/datagenerator.Command"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "successful"
                    },
                    "400": {
                        "description": "{\"message\":\"Error message\"}",
                        "schema": {}
                    },
                    "404": {
                        "description": "{\"message\":\"Error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"message\":\"Error message\"}",
                        "schema": {}
                    }
                }
            }
        },
...
        "/api/v3/resource": {
            "post": {
                "description": "Multiple metric values can be ingested at once, origin can be optionally specified in microseconds",
                "tags": [
                    "hedge-device-virtual"
                ],
                "summary": "ingest the specified machine or sensor data for the device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Name",
                        "name": "deviceName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ReadingData",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/driver.ReadingData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "{\"message\":\"Error message\"}",
                        "schema": {}
                    },
                    "404": {
                        "description": "{\"message\":\"Error message\"}",
                        "schema": {}
                    },
                    "500": {
                        "description": "{\"message\":\"Error message\"}",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "datagenerator.Command": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "driver.Reading": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "origin": {
                    "type": "integer"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "value": {}
            }
        },
        "driver.ReadingData": {
            "type": "object",
            "properties": {
                "readings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/driver.Reading"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "security": [
        {
            "BasicAuth": []
        }
    ]
}